requirements.txt
aiogram
python-dotenv
aiohttp
gigachat